refactor #97: 컨트롤러 헤더 설정
